xform: move limit.opt functions to limit_funcs.go
Release note: None
